IS-2072: Rename web layer to api
