+++ src/main/kotlin/no/nav/syfo/App.kt
@@ -5,10 +5,10 @@ import io.ktor.server.application.*
 import io.ktor.server.config.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
+import no.nav.syfo.api.apiModule
 import no.nav.syfo.infrastructure.azuread.AzureAdClient
 import no.nav.syfo.infrastructure.veiledertilgang.VeilederTilgangskontrollClient
 import no.nav.syfo.infrastructure.wellknown.getWellKnown
-import no.nav.syfo.web.apiModule
 import org.slf4j.LoggerFactory
 import java.util.concurrent.TimeUnit
