+++ src/main/kotlin/no/nav/syfo/infrastructure/veiledertilgang/VeilederTilgangskontrollClient.kt
@@ -8,6 +8,8 @@ import io.ktor.client.statement.*
 import io.ktor.http.*
 import io.micrometer.core.instrument.Counter
 import net.logstash.logback.argument.StructuredArguments
+import no.nav.syfo.api.ClientEnvironment
+import no.nav.syfo.api.httpClientDefault
 import no.nav.syfo.domain.PersonIdent
 import no.nav.syfo.infrastructure.azuread.AzureAdClient
 import no.nav.syfo.infrastructure.metric.METRICS_NS
@@ -15,8 +17,6 @@ import no.nav.syfo.infrastructure.metric.METRICS_REGISTRY
 import no.nav.syfo.util.NAV_CALL_ID_HEADER
 import no.nav.syfo.util.NAV_PERSONIDENT_HEADER
 import no.nav.syfo.util.bearerHeader
-import no.nav.syfo.web.ClientEnvironment
-import no.nav.syfo.web.httpClientDefault
 import org.slf4j.LoggerFactory
 
 class VeilederTilgangskontrollClient(
