+++ src/main/kotlin/no/nav/syfo/infrastructure/wellknown/WellKnownClient.kt
@@ -3,7 +3,7 @@ package no.nav.syfo.infrastructure.wellknown
 import io.ktor.client.call.*
 import io.ktor.client.request.*
 import kotlinx.coroutines.runBlocking
-import no.nav.syfo.web.httpClientProxy
+import no.nav.syfo.api.httpClientProxy
 
 fun getWellKnown(wellKnownUrl: String): WellKnown =
     runBlocking {
